+++ src/app/(main)/setting/terms/page.tsx
@@ -1,4 +1,9 @@
+import type { Metadata } from "next";
 import SettingsDetailLayout from "@/components/base-ui/Settings/SettingsDetailLayout";
+
+export const metadata: Metadata = {
+  title: "이용약관",
+};
 import TermItem from "@/components/base-ui/Settings/Terms/TermItem";
 import { TERMS_DATA } from "@/constants/setting/terms";
